fix: resolve dashboard lint failures blocking CI on work-state dedupe PR
