+++ ui/src/features/factory-graph-editor/lib/factory-graph-editor-connections.ts
@@ -2,16 +2,17 @@ import type { WorkstationProgressOutcomeRouteContext } from "../../current-facto
 import { workstationSupportsProgressOutcomeRoutes } from "../../current-factory-definition/lib/workstation-progress-outcome-routes";
 import { workstationRequiresWorkerAssignment } from "../../current-factory-definition/lib/workstation-worker-assignment";
 import { getFactoryGraphEditorMessages } from "../messages/editor";
-import type {
-  FactoryGraphDraft,
-  FactoryGraphDraftEdgeChange,
-  FactoryGraphEdge,
-  FactoryGraphNode,
-  FactoryGraphNodeKind,
-  FactoryGraphTopology,
-  FactoryWorkstation,
+import {
+  type FactoryGraphDraft,
+  type FactoryGraphDraftEdgeChange,
+  type FactoryGraphEdge,
+  type FactoryGraphNode,
+  type FactoryGraphNodeKind,
+  type FactoryGraphTopology,
+  type FactoryWorkstation,
+  appendUniqueEdgeChange,
+  edgeChangeId,
 } from "./factory-graph-draft-types";
-import { edgeChangeId } from "./factory-graph-draft-types";
 import { PROGRESS_OUTCOME_SOURCE_ANCHOR_IDS } from "./factory-graph-progress-outcome-connection-anchors";
 
 export {
@@ -471,16 +472,5 @@ export function buildFactoryGraphConnectionNotice(options: {
     messages.connectionAnchorLabel(targetAnchor.id),
     options.targetNode.label,
   );
-}
-
-function appendUniqueEdgeChange(
-  edges: FactoryGraphDraftEdgeChange[],
-  edgeChange: FactoryGraphDraftEdgeChange,
-) {
-  const nextEdgeId = edgeChangeId(edgeChange);
-  if (edges.some((entry) => edgeChangeId(entry) === nextEdgeId)) {
-    return edges;
-  }
-  return [...edges, edgeChange];
 }
